untabify all the files.
There is a problem of configuration somewhere.
Let say that the indentation is 4 *spaces* per TAB.

And we do not store TAB but spaces.
M-x whitespace-mode allows to display the difference.

i.e. (setq-default indent-tabs-mode nil) normally does the job.
Otherwise, select a region and apply M-x untabify (or by a hook).